Add support for functional components
I had to add a requirements of them to be in pascal case to distinguish
them from regular tags. It's not ideal, but I cannot find a better solution.
Macros are executed in compile time so I cannot keep a list of functions that
are tagged with #[component] to look up it later when html! macro is invoked.

Other libs and frameworks also do something similar it seems, although
I didn't looked into their implementation.

Whith this implemented, MVP is done. Time to think about docs, examples and
a crates.io release