Add support for iteration using map
Wow, this was not an easy one, and it's still hacky. Like, I check specifically
that the expression has a call to a method named "map" and then treat
it differently.

Also I had to make html! macro to always return a NodeCollection and
I don't think it's a good thing. Ability to create a single Node
via the macro looks like a nice feature to have.

I'm not sure if this feature will stay here long term