Implement slots arguments in component
Currently lilmited to only one. Not sure about allowing more and I don't
think it's a priority right now.

Slot argument needs to be put in double braces like so `{{ slot }}`
to disable html escaping, otherwise it won't work as expected.
I'm not very fond of it, but not sure what could be the alternative.

Closes #10