Impelement bare text in text nodes
In Rust 1.88 Span API was stabilized so it's now possible without
losing the space characters.

Unfortunately, there is no way to accept `'` in there using
2021 or 2024 edition of Rust. I don't think I want to downgrade to
2018 so this part will have to be sacraficed. Percy can handle it,
but it is stuck on 2018.

I still haven't figured what to do with the spaces around expressions,
so this is very much a work in progress and will to go to a branch first